move viewmodel to core
