+++ src/Core/ViewModel/BaseViewModel.js
